Break up riginfo and bone surgery
