+++ blender_source/MH_Community/operators/amputateface.py
@@ -2,6 +2,7 @@
 # -*- coding: utf-8 -*-
 
 import bpy
+from ..rig import RigInfo
 
 class AmputateFaceOperator(bpy.types.Operator):
     """Remove face bones, and assign their weights to head bone"""
@@ -10,8 +11,6 @@ class AmputateFaceOperator(bpy.types.Operator):
     bl_options = {'REGISTER', 'UNDO'}
 
     def execute(self, context):
-        from ..rig_info import RigInfo
-        from ..bone_surgery import BoneSurgery
         armature = context.object
 
         rigInfo = RigInfo.determineRig(armature)
@@ -31,7 +30,6 @@ class AmputateFaceOperator(bpy.types.Operator):
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
     @classmethod
     def poll(cls, context):
-        from ..rig_info import RigInfo
         ob = context.object
         if ob is None or ob.type != 'ARMATURE': return False
