+++ blender_source/MH_Community/rig/__init__.py
@@ -1,0 +1,18 @@
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+from .riginfo import RigInfo
+from .cmuriginfo import CMURigInfo
+from .defaultriginfo import DefaultRigInfo
+from .gameriginfo import GameRigInfo
+from .kinect2riginfo import Kinect2RigInfo
+from .bonesurgery import BoneSurgery
+
+__all__ = [
+    RigInfo,
+    CMURigInfo,
+    DefaultRigInfo,
+    GameRigInfo,
+    Kinect2RigInfo,
+    BoneSurgery
+]
