+++ blender_source/MH_Community/snap_on_ik_rig.py
@@ -1,6 +1,8 @@
 from .rig_info import *
 
 import bpy
+from .rig import DefaultRigInfo
+
 #===============================================================================
 class IkRig():
     def __init__(self, rigInfo):
